src/backend/pyproject.toml
[project]
name = "companion-ai-backend"
version = "0.1.0"
description = "Backend para Companion AI — agente conversacional para adultos mayores"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "langgraph>=0.2",
    "langchain-google-genai>=2.1",
    "google-generativeai>=0.8",
    "elevenlabs>=1.9",
    "pymongo[srv]>=4.9",
    "motor>=3.6",
    "python-multipart>=0.0.12",
    "pydantic-settings>=2.5",
    "httpx>=0.27",
]
